requirements.txt
requests~=2.27.1
tornado~=6.1
enums~=0.0.2
PyYAML~=6.0
pip~=21.3.1
wheel~=0.37.1
certifi~=2021.10.8
idna~=3.3
urllib3~=1.26.8
setuptools~=60.0.5
coloredlogs~=15.0.1
concurrent-log-handler~=0.9.19
sqlalchemy-utils~=0.38.2
SQLAlchemy~=1.4.31
alembic~=1.7.6
psycopg2~=2.9.1
psycopg2-binary~=2.9.1
cryptography~=36.0.1